refactor(networks): inject Connect through the shared module API
