+++ networks/ethereum/network-ethereum-suite-common/src/createEthereumSuiteCommonNetworkModule.ts
@@ -6,17 +6,30 @@ import {
 import type { SuiteCommonNetworkModule } from '@trezor/network-module-suite-common-types';
 
 import { ethereumValidator } from './addressValidator/ethereumAddressValidator';
-import { ethereumNamedAddressResolver } from './namedAddress/ethereumNamedAddressResolver';
+import {
+    type EthereumNamedAddressResolverCompositionRootDeps,
+    createEthereumNamedAddressResolverCompositionRoot,
+} from './namedAddress/createEthereumNamedAddressResolverCompositionRoot';
 import { getNetworkConfig } from './networkConfig';
+
+type EthereumSuiteCommonNetworkModuleDeps = EthereumNamedAddressResolverCompositionRootDeps;
 
 export type EthereumNetworkSuiteCommonNetworkModule =
     SuiteCommonNetworkModule<EthereumNetworkSymbol>;
 
-export const createEthereumSuiteCommonNetworkModule =
-    (): EthereumNetworkSuiteCommonNetworkModule => ({
+type EthereumSuiteCommonNetworkModule = EthereumNetworkSuiteCommonNetworkModule;
+
+export const createEthereumSuiteCommonNetworkModule = (
+    deps: EthereumSuiteCommonNetworkModuleDeps,
+): EthereumSuiteCommonNetworkModule => {
+    const { ethereumNamedAddressResolver } =
+        createEthereumNamedAddressResolverCompositionRoot(deps);
+
+    return {
         addressValidator: ethereumValidator,
         namedAddressResolver: ethereumNamedAddressResolver,
         getSupportedNetworks: () => supportedEthereumNetworks,
         isSupportedNetwork: isSupportedEthereumNetwork,
         getNetworkConfig,
-    });
+    };
+};
