+++ networks/ethereum/network-ethereum-suite-common/src/index.ts
@@ -1,5 +1,5 @@
-export { createEthereumSuiteCommonNetworkModule } from './EthereumNetworkSuiteCommonNetworkModule';
-export type { EthereumNetworkSuiteCommonNetworkModule } from './EthereumNetworkSuiteCommonNetworkModule';
+export { createEthereumSuiteCommonNetworkModule } from './createEthereumSuiteCommonNetworkModule';
+export type { EthereumNetworkSuiteCommonNetworkModule } from './createEthereumSuiteCommonNetworkModule';
 
 // These exports are temporary migration aids. Once network modularization is complete,
 // wrapped-native token configuration shall remain private to the Ethereum network module.
