+++ networks/ethereum/network-ethereum-suite-common/src/namedAddress/ReverseResolveAddress.ts
@@ -1,0 +1,11 @@
+import type { EthereumNetworkSymbol } from '@trezor/network-ethereum/constants';
+
+/** @serviceContract */
+export type ReverseResolveAddress = (
+    address: string,
+    symbol: EthereumNetworkSymbol,
+) => Promise<string | null>;
+
+export type ReverseResolveAddressDep = {
+    reverseResolveAddress: ReverseResolveAddress;
+};
