+++ networks/ethereum/network-ethereum-suite-common/src/namedAddress/createEthereumNamedAddressResolverCompositionRoot.test.ts
@@ -1,17 +1,21 @@
 import { encodeErrorResult, parseAbi } from 'viem';
 
-import { resolveNamedAddress } from './resolveNamedAddress';
+import {
+    type EthereumNamedAddressResolverCompositionRootDeps,
+    createEthereumNamedAddressResolverCompositionRoot,
+} from './createEthereumNamedAddressResolverCompositionRoot';
 
 const mockBlockchainEvmRpcCall = jest.fn();
 const mockGetAccountInfo = jest.fn();
 
-jest.mock('@trezor/connect', () => ({
-    __esModule: true,
-    default: {
-        blockchainEvmRpcCall: (...args: unknown[]) => mockBlockchainEvmRpcCall(...args),
-        getAccountInfo: (...args: unknown[]) => mockGetAccountInfo(...args),
-    },
-}));
+const deps: EthereumNamedAddressResolverCompositionRootDeps = {
+    getTrezorConnect: () => ({
+        blockchainEvmRpcCall: mockBlockchainEvmRpcCall,
+        getAccountInfo: mockGetAccountInfo,
+    }),
+};
+const { ethereumNamedAddressResolver } = createEthereumNamedAddressResolverCompositionRoot(deps);
+const { resolveNamedAddress } = ethereumNamedAddressResolver;
 
 /**
  * Guards the number of requests a single resolution costs, through the real resolver and its
