+++ networks/ethereum/network-ethereum-suite-common/tsconfig.json
@@ -5,11 +5,13 @@
         {
             "path": "../../../suite-common/calldata"
         },
-        { "path": "../../../packages/connect" },
         { "path": "../network-ethereum" },
         {
             "path": "../../network-module/network-module-suite-common-types"
         },
-        { "path": "../../../packages/utils" }
+        { "path": "../../../packages/utils" },
+        {
+            "path": "../../../suite-common/dependency-injection"
+        }
     ]
 }
