+++ networks/ethereum/network-ethereum-suite-common/tsconfig.lib.json
@@ -6,11 +6,13 @@
         {
             "path": "../../../suite-common/calldata"
         },
-        { "path": "../../../packages/connect" },
         { "path": "../network-ethereum" },
         {
             "path": "../../network-module/network-module-suite-common-types"
         },
-        { "path": "../../../packages/utils" }
+        { "path": "../../../packages/utils" },
+        {
+            "path": "../../../suite-common/dependency-injection"
+        }
     ]
 }
