+++ networks/network-module/network-module-suite-common-types/mocks/index.ts
@@ -1,0 +1,1 @@
+export { mockGetTrezorConnect } from './mockGetTrezorConnect';
