+++ networks/network-module/network-module-suite-common-types/package.json
@@ -11,6 +11,7 @@
         "type-check": "yarn g:tsc --build"
     },
     "dependencies": {
+        "@trezor/connect-common": "workspace:*",
         "@trezor/type-utils": "workspace:*"
     }
 }
