+++ networks/network-module/network-module-suite-common-types/src/GetTrezorConnect.ts
@@ -1,0 +1,14 @@
+import type { TrezorConnectCallable } from '@trezor/connect-common';
+
+/**
+ * Reads Connect at call time because legacy initialization replaces its methods.
+ * The generic limits each consumer to the Connect methods it actually requires.
+ */
+export type GetTrezorConnect<TMethod extends keyof TrezorConnectCallable> = () => Pick<
+    TrezorConnectCallable,
+    TMethod
+>;
+
+export type GetTrezorConnectDep<TMethod extends keyof TrezorConnectCallable> = {
+    getTrezorConnect: GetTrezorConnect<TMethod>;
+};
