+++ networks/network-module/network-module-suite-common-types/src/NetworkSuiteCommonModuleApi.ts
@@ -1,0 +1,9 @@
+import type { GetTrezorConnectDep } from './GetTrezorConnect';
+
+/**
+ * Application capabilities injected into suite-common network modules.
+ * Apps choose their Connect implementation; modules must not import its runtime themselves.
+ */
+export type NetworkSuiteCommonModuleApi = GetTrezorConnectDep<
+    'getAccountInfo' | 'blockchainEvmRpcCall'
+>;
