+++ networks/network-module/network-module-suite-common-types/src/index.ts
@@ -1,6 +1,8 @@
 export { addressType } from './AddressValidator';
 export type { AddressType, AddressValidator } from './AddressValidator';
 export type { NamedAddressResolver } from './NamedAddressResolver';
+export type { NetworkSuiteCommonModuleApi } from './NetworkSuiteCommonModuleApi';
+export type { GetTrezorConnect, GetTrezorConnectDep } from './GetTrezorConnect';
 export { asProtocol } from './Protocol';
 export type { Protocol } from './Protocol';
 export type {
