+++ networks/network-module/network-module-suite-common-types/tsconfig.json
@@ -2,6 +2,9 @@
     "extends": "../../../tsconfig.base.json",
     "compilerOptions": { "outDir": "libDev" },
     "references": [
+        {
+            "path": "../../../packages/connect-common"
+        },
         { "path": "../../../packages/type-utils" }
     ]
 }
