+++ networks/network-module/network-module-suite-common-types/tsconfig.lib.json
@@ -6,6 +6,9 @@
     },
     "include": ["./src"],
     "references": [
+        {
+            "path": "../../../packages/connect-common"
+        },
         { "path": "../../../packages/type-utils" }
     ]
 }
