+++ suite-common/address/src/autocorrectAddress.test.ts
@@ -3,6 +3,7 @@ import {
     createNetworksCompositionRoot,
 } from '@suite-common/networks';
 import { asNetworkSymbol } from '@suite-common/wallet-config';
+import { mockGetTrezorConnect } from '@trezor/network-module-suite-common-types/mocks';
 
 import { createAddressValidator } from './AddressValidator';
 import { autocorrectAddress } from './autocorrectAddress';
@@ -12,7 +13,9 @@ const bchSymbol = asNetworkSymbol('bch');
 const ethSymbol = asNetworkSymbol('eth');
 
 describe('autocorrectAddress', () => {
-    const networkModules = createNetworksCompositionRoot();
+    const networkModules = createNetworksCompositionRoot({
+        getTrezorConnect: mockGetTrezorConnect,
+    });
     const networkModuleRepository = createNetworkModuleRepository({ networkModules });
     const addressValidator = createAddressValidator({
         networkModuleRepository,
