+++ suite-common/address/src/isAddressDeprecated.test.ts
@@ -3,6 +3,7 @@ import {
     createNetworksCompositionRoot,
 } from '@suite-common/networks';
 import { asNetworkSymbol } from '@suite-common/wallet-config';
+import { mockGetTrezorConnect } from '@trezor/network-module-suite-common-types/mocks';
 
 import { createAddressValidator } from './AddressValidator';
 import { isAddressDeprecated } from './isAddressDeprecated';
@@ -13,7 +14,9 @@ const bchSymbol = asNetworkSymbol('bch');
 // https://litecoin-project.github.io/p2sh-convert/
 // https://cashaddr.bitcoincash.org/
 describe('isAddressDeprecated', () => {
-    const networkModules = createNetworksCompositionRoot();
+    const networkModules = createNetworksCompositionRoot({
+        getTrezorConnect: mockGetTrezorConnect,
+    });
     const networkModuleRepository = createNetworkModuleRepository({ networkModules });
     const addressValidator = createAddressValidator({
         networkModuleRepository,
