+++ suite-common/networks/src/createNetworksCompositionRoot.ts
@@ -1,6 +1,7 @@
 import { createBitcoinSuiteCommonNetworkModule } from '@trezor/network-bitcoin-suite-common';
 import { createCardanoSuiteCommonNetworkModule } from '@trezor/network-cardano-suite-common';
 import { createEthereumSuiteCommonNetworkModule } from '@trezor/network-ethereum-suite-common';
+import type { NetworkSuiteCommonModuleApi } from '@trezor/network-module-suite-common-types';
 import { createRippleSuiteCommonNetworkModule } from '@trezor/network-ripple-suite-common';
 import { createSolanaSuiteCommonNetworkModule } from '@trezor/network-solana-suite-common';
 import { createStellarSuiteCommonNetworkModule } from '@trezor/network-stellar-suite-common';
@@ -8,13 +9,17 @@ import { createTronSuiteCommonNetworkModule } from '@trezor/network-tron-suite-c
 
 import { type NetworkModules } from './NetworkModules';
 
-export const createNetworksCompositionRoot = (): NetworkModules => {
+type NetworksCompositionRootDeps = NetworkSuiteCommonModuleApi;
+
+export const createNetworksCompositionRoot = (
+    deps: NetworksCompositionRootDeps,
+): NetworkModules => {
     // When adding a new Network Module, you have to
     //    1. register it here to have the runtime object for DI
     //    2. and in the `NetworkModules` to have static typings right
     const networkModules: NetworkModules = {
         bitcoin: createBitcoinSuiteCommonNetworkModule(),
-        ethereum: createEthereumSuiteCommonNetworkModule(),
+        ethereum: createEthereumSuiteCommonNetworkModule(deps),
         ripple: createRippleSuiteCommonNetworkModule(),
         cardano: createCardanoSuiteCommonNetworkModule(),
         solana: createSolanaSuiteCommonNetworkModule(),
