+++ suite/e2e/support/testExtends/customMatchers.ts
@@ -11,6 +11,7 @@ import {
     createNetworksCompositionRoot,
 } from '@suite-common/networks';
 import { type Account } from '@suite-common/wallet-types';
+import { mockGetTrezorConnect } from '@trezor/network-module-suite-common-types/mocks';
 import { Model } from '@trezor/trezor-user-env-link';
 import { getIndexOrThrow } from '@trezor/utils';
 
@@ -25,7 +26,7 @@ const DISPLAY_CHAR_LIMIT_T3T1 = 18;
 const STRING_UP_TO_T3T1_DISPLAY_LIMIT = new RegExp(`.{1,${DISPLAY_CHAR_LIMIT_T3T1}}`, 'g');
 const intlEn = createIntl({ locale: 'en', messages: {} }, createIntlCache());
 
-const networkModules = createNetworksCompositionRoot();
+const networkModules = createNetworksCompositionRoot({ getTrezorConnect: mockGetTrezorConnect });
 const networkModuleRepository = createNetworkModuleRepository({ networkModules });
 const addressValidator = createAddressValidator({ networkModuleRepository });
